Apply GoF's state pattern to remove conditional logic
While our first tought could be to use polymorphism, in this case it
doesn't work because a movie could change its price_code during its
lifetime.

In this case it is much better to use the GoF's state/strategy pattern
and having a price attribute, containing the reference to the current
state.

Why should we use state instead of strategy? Does the price class
represent an algorithm for calculating the price (in which case it would
be better to call it Pricer), or does it represent a state of the movie
(is this movie a new release)? The choice of the pattern (and name)
reflects how you want to think about the structure. At the moment it
will be a state of a movie, but if we later decide a strategy
communicates the intention better, we can always refactor it by changing
the names.